Use the whole serial buffer and drop old data on overflow
The serial buffer was 1k, but an un8 was used as index. By changing to
unsigned int, the whole buffer is used.

Increase buffer size to same size as the serial thread (2k)

In case of a receive overflow, drop the old buffer in
favor of new data, in an attempt to sync again.